+++ Assets/LDtkUnity/Editor/ScriptedImporter/LDtkLevelImporter.cs
@@ -103,16 +103,16 @@ namespace LDtkUnity.Editor
 
         private void BuildLevel()
         {
-            LDtkPostProcessorCache postProcess = new LDtkPostProcessorCache();
+            LDtkAssetProcessorActionCache assetProcess = new LDtkAssetProcessorActionCache();
 
-            LDtkBuilderLevel levelBuilder = new LDtkBuilderLevel(_projectImporter, _projectJson, WorldLayout.Free, _levelJson, postProcess, this);
+            LDtkBuilderLevel levelBuilder = new LDtkBuilderLevel(_projectImporter, _projectJson, WorldLayout.Free, _levelJson, assetProcess, this);
             GameObject levelRoot = levelBuilder.StubGameObject();
             
             Profiler.BeginSample($"BuildSeparateLevel {_levelJson.Identifier}");
             levelBuilder.BuildLevel();
             Profiler.EndSample();
             
-            postProcess.PostProcess();
+            assetProcess.Process();
 
             ImportContext.AddObjectToAsset("levelRoot", levelRoot, _icon);
             ImportContext.SetMainObject(levelRoot);