+++ Assets/LDtkUnity/Runtime/Data/Schema/Instance/FieldInstance.cs
@@ -26,15 +26,7 @@ namespace LDtkUnity
         public string Type { get; set; }
 
         /// <summary>
-        /// Actual value of the field instance. The value type varies, depending on `__type`:<br/>
-        /// - For **classic types** (ie. Integer, Float, Boolean, String, Text and FilePath), you
-        /// just get the actual value with the expected type.<br/>   - For **Color**, the value is an
-        /// hexadecimal string using "#rrggbb" format.<br/>   - For **Enum**, the value is a String
-        /// representing the selected enum value.<br/>   - For **Point**, the value is a
-        /// [GridPoint](#ldtk-GridPoint) object.<br/>   - For **Tile**, the value is a
-        /// [TilesetRect](#ldtk-TilesetRect) object.<br/>   - For **EntityRef**, the value is an
-        /// [EntityReferenceInfos](#ldtk-EntityReferenceInfos) object.<br/><br/>  If the field is an
-        /// array, then this `__value` will also be a JSON array.
+        /// Actual value of the field instance.
         /// </summary>
         [DataMember(Name = "__value")]
         public object Value { get; set; }