+++ Assets/LDtkUnity/CHANGELOG.md
@@ -1,6 +1,7 @@
 # 4.1.1
 ###### January 1, 2023
 - Updated json support & samples for [LDtk 1.5](https://ldtk.io/json/#changes;1.5.0)
+- Added LDtkPreprocessor: Use to read/modify any json data before hierarchies are created
 
 # 4.1.0
 ###### October 23, 2023